pyproject.toml
[project]
name = "labelstudio-tools"
version = "0.1.0"
description = "Working with Labelstudio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepdiff>=8.4.2",
    "deprecated>=1.2.18",
    "httpx>=0.28.1",
    "irrcac>=0.4.4",
    "jsonpath-ng>=1.7.0",
    "lxml>=5.3.2",
    "orjson>=3.10.16",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "pystache>=0.6.8",
    "tqdm>=4.67.1",
    "typer>=0.15.2",
]

[project.optional-dependencies]
plot = [
    "matplotlib>=3.10.1",
]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.6",
]
